requirements.txt
tqdm==4.65.0
Pillow
opencv-python
imageio
matplotlib
kornia
natsort
pyyaml
wandb
lpips
open3d==0.16.0
torchmetrics
cyclonedds
pytorch-msssim
open-clip-torch
plyfile==0.8.1
git+https://github.com/JonathonLuiten/diff-gaussian-rasterization-w-depth.git@cb65e4b86bc3bd8ed42174b72a62e8d3a3a71110